cleaning up the models
